requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic[email]==2.11.7
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.34
pymysql==1.1.0
alembic==1.13.1

# Security
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# HTTP & Networking
requests==2.32.5
httpx==0.27.0

# Configuration & Environment
python-dotenv==1.1.1

# Validation & Phone Numbers
phonenumbers==8.13.27

# Caching (Optional)
redis==5.0.1

# Development & Testing
pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio==0.23.6

# Code Quality
black==24.4.2
flake8==7.0.0
isort==5.13.2
mypy==1.10.0
bandit==1.7.5
safety==3.2.0

# Monitoring & Logging
structlog==24.1.0

# Production
gunicorn==22.0.0
